Remove warning at top of git web pages.
Although the documentation is still a work in progress, with various
forms of git usage (e.g. contrib/gcc-git-customization.sh) not yet
documented and development branches not yet listed on git.html as they
are on svn.html, the warning about the conversion still being in
progress is no longer accurate, so remove it.