Copy list of development branches from svn.html into git.html.
This patch makes a start on making the branch information from
svn.html available in git.html.  The list is copied, some notes added
on devel/, refs/dead/heads/ and refs/vendors/ naming, and inactive
branches from the general section (ones not present in
refs/heads/devel/, automatically moved to refs/dead/heads/ as part of
the conversion, because of lack of commits since at least December
2017) are moved into the section of inactive branches that might or
might not have been merged.  Significant further work on the list is
needed to ensure that all inactive branches are moved to inactive
sections, that active branches are properly listed in active sections,
and that it is clear for each branch where it appears in git
(including vendor branches, and git-mirror-only branches that are now
in refs/git-old/heads).