document scripts to access personal and vendor spaces
This patch documents some of the scripts that I've published for
managing the personal and vendor spaces on the server.  It also covers
some of the other features that those scripts enable, so that it's all
in one place.  This is a complete rewrite of the material I had
written previously since before we did not have these scripts to make
use of.

I've not filled in the documentation for gcc-descr and gcc-undescr,
Jakub has agreed to provide that at a later date.