[wwwdocs] Move arm section under aarch64
We logically grouped the aarch64 and arm entries as they share some content but
then the alphabetic order came in and put amdgcn and arc between the two.
I'm taking the liberty of moving arm under aarch64 and adding a small clarification
that the recently added CDE intrinsics implement a beta version of the ACLE specification.