Correct status and macro for P1907R1
As noted in PR c++/94765 floating point types are not supported as
non-type template args.

Marek added the P1907R1 paper in commit d59a823fb but didn't give it a
distinct entries for the "Supported in GCC?" and "SD-6 Feature Test"
columns. This patch keeps it grouped with P0732R2 but in its own row.

The table uses the nth-child CSS pseudo-class for formatting, but
rowspan attributes mess that up by altering the number of children in
the row. This patches uses (non-displayed) <template> elements to
preserve the ordering for nth-child, so that the table cells are aligned
consistently. Those elements have no other effect, and no effect at all
for non-graphical browsers that ignore the CSS formatting.