Don't list RMS as member of the Steering Committee
Per https://gcc.gnu.org/pipermail/gcc/2021-March/235245.html

"In 2012 RMS was added to the GCC Steering Committee web page based
on his role in the GNU Project, though his role as a member of the
Steering Committee has been ambiguous and he was not a member of
the Steering Committee when EGCS became GCC[1].  We no longer feel
that this listing serves the best interests of the GCC developer
and user community.  Therefore, we are removing him from the page.

GCC supports the principles of Free Software and has remained
aligned with the GNU Project since EGCS became GCC, but effectively
has continued to operate as an autonomous project.

The GCC Steering Committee is committed to providing a friendly, safe
and welcoming environment for all, regardless of gender identity and
expression, sexual orientation, disabilities, neurodiversity, physical
appearance, body size, ethnicity, nationality, race, age, religion, or
similar personal characteristics.

- The GCC Steering Committee

[1] https://static.lwn.net/1999/0429/a/gcc.html"