lists: Fix thinko
Brown paper bag time.  The List-Id: should look like a hostname, not
like an email address.  Somehow I put in an at-sign when changing my
gcc-patches example to the match-all example we have here.

Note that the "." in the procmail match are RE wildcards btw.  This is
common practice in procmailrcs, although not necessarily a good idea :-)