wwwdocs: Clarify DCO name/identity and (anonymous) pseudonym policy
Adjust the DCO text to match the broader community usage and
clarifications around the use of real names, known identities and
(anonymous) pseudonyms.

These changes clarify what was meant by "real name" and that it is not
required to be a "legal name" or any other stronger requirement than a
known identity and can be contacted to discuss the contribution as
adopted by other communities like the linux kernel, elfutils, cncf and
gentoo.

Also explain that the FSF assignment policy might be more appropriate
when wanting to contribute using an anonymous pseudonym or when
needing an explicit company disclaimer.