wwwdocs: add a Python postprocessing script
The heading elements in our website contain "id" information,
but currently to find them you to look at the page source,
whereas in the generated HTML for the manual we have e.g.:

  <a class="copiable-link" href="#index-mabi-1"> &para;</a>

which shows up nicely in the browser in e.g.
  https://gcc.gnu.org/onlinedocs/gcc/ARM-Options.html
as a pilcrow character when you hover over the link, which
you can then use to copy the URL to the clipboard.

It's *very* helpful to have easily shareable links to within pages.

The attached patch adds a postprocessing step to "bin" that
turns e.g.
  <h1 id="ID">TEXT</h1>
to:
  <h1 id="ID"><a href="#ID">TEXT</a></h1>

which makes it very easy to copy links in the generated website.

Hook this script into our general preprocessing script for web pages.